+++ frontend/src/pages/Dashboard/Dashboard.tsx
@@ -6,12 +6,12 @@ import {
   Redirect,
 } from 'react-router-dom';
 import Sidebar from '../../components/Sidebar/Sidebar';
-import Drivers from './Drivers';
+import Employees from './Employees';
 import Students from './Students';
 import Home from './Home';
 import Locations from './Locations';
 import Settings from './Settings';
-import DriverDetail from '../../components/UserDetail/DriverDetail';
+import EmployeeDetail from '../../components/UserDetail/EmployeeDetail';
 import RiderDetail from '../../components/UserDetail/RiderDetail';
 import ExportPreview from '../../components/ExportPreview/ExportPreview';
 import DateContext from '../../context/date';
@@ -38,8 +38,8 @@ const Dashboard = () => {
               path="/employees"
               render={({ match: { url } }) => (
                 <>
-                  <Route path={`${url}/`} component={Drivers} exact />
-                  <Route path={`${url}/employee`} component={DriverDetail} />
+                  <Route path={`${url}/`} component={Employees} exact />
+                  <Route path={`${url}/employee`} component={EmployeeDetail} />
                 </>
               )}
             />
